quantum_os/requirements.txt
# Quantum OS Requirements
# Core quantum computing frameworks

# Google Cirq (for Willow backend)
cirq>=1.3.0
cirq-google>=1.3.0

# IBM Qiskit (for Brisbane/Torino backends)
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-ibm-runtime>=0.17.0
qiskit-ibm-provider>=0.8.0

# TensorFlow Quantum (hybrid quantum-classical ML)
tensorflow>=2.15.0
tensorflow-quantum>=0.7.3

# PyTorch for hybrid models
torch>=2.1.0
torchvision>=0.16.0

# Quantum machine learning
pennylane>=0.33.0
pennylane-qiskit>=0.33.0
pennylane-cirq>=0.33.0

# Classical ML frameworks
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0

# GPU acceleration
cupy-cuda12x>=12.3.0  # For CUDA 12.x
pycuda>=2022.2

# Distributed computing
dask[complete]>=2023.10.0
ray[default]>=2.8.0
mpi4py>=3.1.5

# Networking and API
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Security and cryptography
cryptography>=41.0.5
pynacl>=1.5.0
pyarmor>=8.4.0  # For code obfuscation

# Configuration and logging
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2
hydra-core>=1.3.2

# Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.6

# Cloud providers
google-cloud-quantum>=0.1.0
boto3>=1.29.0  # For AWS Braket (future support)

# Development tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
mypy>=1.7.0
pylint>=3.0.2

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Database for quantum job tracking
sqlalchemy>=2.0.23
redis>=5.0.1

# Optimization
ortools>=9.7.0
cvxpy>=1.4.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.2
click>=8.1.7
